# src/main/java/com/eagle/api/dto/CreateBankAccountRequest.java
package com.eagle.api.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class CreateBankAccountRequest {
    @NotBlank
    private String name;
    @NotBlank
    private String accountType;
}