# src/main/java/com/eagle/api/dto/LoginRequest.java
package com.eagle.api.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {
    @NotBlank
    private String username;
    @NotBlank
    private String password;
}